+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Add the `null` combinator.
 - Add `obj0` (matches any object) and `empty` (matches the empty object)
   combinators.
+- Add the `union` combinator.
 
 # 2.0.0 (2024-01-12)